requirements.txt
streamlit==1.25.0
requests==2.31.0
pandas==2.0.3
openai==0.27.8
python-dotenv==1.0.0
plotly==5.15.0